Add CheckedGroupShoppingListItemModalContainer.tsx and EditGroupShoppingListItemModalContainer.tsx
