+++ src/containers/shoppingList/modules/ListItem/shoppingListItemComponent/ShoppingListItemComponentContainer.tsx
